+++ src/server.ts
@@ -16,11 +16,10 @@ export class Server extends Agent<IpcMain> {
 
   /**
    * Sends a message to the renderer process.
-   * @param channel The channel to use for sending the data
    * @param message The message to send
    */
-  protected send (channel: string, message: Message): void {
-    this.webContents.send(channel, message)
+  protected send (message: Message): void {
+    this.webContents.send(Server.ipcChannel, message)
   }
 
 }